BUG#21031993 : Fix mysqluserclone failing with --include-global-priviliges
When mysqluserclone utility is executed with --include-global-priviliges it tries
to grant privilige to the source user on the destination server instead of the
destination user. This was oaccuring due to incorrectly replacing source host
in the grant query.

No test case was added as the patch fixes already broken tests.