BUG#20979041 : mysqlrplsync gives wrong information with server 5.7.7
The mysqlrplsync utility is comparing the 'sys' internal database for 5.7.7
servers if no specific databases or tables are specified. The 'sys' schema
is now installed by default starting for MySQL 5.7.7, but the utility should
ignore it by default, like for other internal database (such as mysql).

This patches fixes this issue by treating the 'sys' database as an internal
one for 5.7.7 server or higher.

In addition, the creation of the replication user on the master for
mysqlreplicate was fixed, ensuring that the user is created with the
specified password and without printing the CREATE USER statement (in
verbose mode).

No new MUT tests was changed or added because existing ones already detected
this issue.