BUG#21110885 : some MUT tests fail with server 5.7.7 due to mask issues
Some MUT tests fail due to mask issues when running with a MySQL 5.7.7 server,
in particular: check_rpl_table, export_gtid, and
rpl_admin_gtid_demote_master_repo_file. The export_gtid test was falling
because the 'sys' database (added as default for 5.7.7) needs to be handled
as an internal database. The rpl_admin_gtid_demote_master_repo_file test was
falling because the default binlog format changed for 5.7.7 and different
GTID sets are generated for STATEMENT and ROW based replication. The
remaining issue were related with expected non-deterministic outputs between
different server versions that needed to be masked.

This patches fixes the mask issues for the referred tests, ensuring that
expected non-deterministic outputs for 5.7.7 servers are correctly handled.