BUG#21327864 : mysqldbimport utility failing with error 1694 with mysql 5.7.7
The mysqldbimport utility was failing with error 1694 for MySQL 5.7.7 when
executing a SET SQL_LOG_BIN statement preceded by a SET GTID_PURGED. This
behavior is due to an issue with 5.7.7 with the SET GTID_PURGED that
incorrectly starts a new transaction.

This patch fixes this issue by improving the code handling the this error,
now always performing a commit when a SET GTID_PURGED is detected and not
only when preceded by a DML command.