BUG#21343759 : Fix my_print_defaults to read .mylogin.cnf on MySQL 5.6.25
As MySQL Server 5.6.25 and 5.7.8 a change to the my_print_defaults tool
was introduced that consist on masking the passwords and making them not
visible as plain text, this causes the retrieval of the password to fail
when mylogin is used with the MySQL Utilities. This fix uses the '--show'
password option introduced in the my_print_defaults tool starting from
the same MySQL Server versions to retrieve the password with the help
of my_print_defaults and adding this option to the version of the server
where the version is5.6.25 or higher but lower to 5.7.0 or 5.7.8 and
higher.

v.2