BUG#21839348 : AttributeError in mysqldbcompare using --all option
Using the same connection information on both --server1 and --server2
with the --all option, it raised an AttributeError in mysqldbcompare.

The connect_servers() function used by compare_all_databases() sets
None for the destination server if the source and destination are the
same server and the unique parameter is False.
It was added a check if the destination server is None to fix this issue.

A MUT test was added for regression.