BUG#21653352 : Utilities fails with Server SQL mode set to ANSI QUOTES
The utilities fail to execute when the sql_mode=ANSI_QUOTES is set on the
MySQL Server. While the bug report mysqlrpladmin failing due to a internal
query that can not be executed when the server is set to
sql_mode=ANSI_QUOTES
because of the use of double quotes. In addition this sql_mode mask the
hashed password with the string '<secret>' causing errors while parsing a
grant statement which limits gathering user's permission and more
utilities
to fail with this mode.

This patch contains the following changes to fix the issues mentioned
above:
-The regular expression to parse the grant statements has been updated to
 support statements with the hashed password masked.
-The queries on common\replication.py has been modified to avoid use of
 double quotes (") and replaced with single quotes (') capable to run
 regardless of the sql_mode set in the server.
-New regular expressions had been added to parse the object identifiers
 for the use of double quotes (").

With this patch the users do not need to specify the sql_mode they use
in the server, this is queried from the servers given to the utility, but
users need to be careful to use the escape sequences when the objects
identifier contains backtick quotes. While a backtick quote needed to be
escape by a second backtick, the ANSI quotes sql_mode does not require such
escape.

Servers with sql_mode set to ANSI_QUOTES or a mixed sql_mode can now
be used  with the following utilities mysqldbcompare, mysqldbcopy,
mysqldbexport, mysqldbimport, mysqldiff, mysqlgrants, and mysqlrpladmin.
The mysqlrplsync requires to all or none of the servers to be set with
theSQL mode set to ANSI_QUOTES.

New MUT tests has been added in this patch to exercise the changes.

v.4