BUG#21572065 : mysqldbcompare raise errors when the index definitions differs
The mysqldbcompare raise exceptions comparing tables when the used indexes
have different column definitions due to the use of the same index
definition to create both compare tables, which can compromise the values of
the indexes to be inserted in the compare tables due to lost of precision or
data truncation. This patch ensures each compare table is created based in
the column definitions from the the table from the indexes are load from.

A mut test was added to exercise the changes.

The BUG#21803368 was marked as duplicated of this bug.
Note added mut test is limited to run in servers prior to 5.7
v.1c