BUG#22608528 : mysqlrplsync does not work with mysql 5.7
The mysqlrplsync utility is not working with MySQL 5.7 if we use it in
a topology where there is a master. The utility fails when trying to obtain
the master UUID from the corresponding variable after locking the table that
is being checked. This issue happens because starting from MySQL 5.7.6 the
the "session_variables" table was moved from the INFORMATION_SCHEMA to the
PERFORMANCE_SCHEMA, and the tables in the INFORMATION_SCHEMA are the only
exception of tables than can be accessed without requiring to be specified
by the LOCK statement.

This patch addresses this issue by retrieving the master UUID before
locking the tables to get the synchronization point.