BUG#22617526 : mysqlfailover and mysqlrpladmin do not work with mysql57
The mysqlfailover and mysqlrpladmin utilities are not working with
mysql5.7 servers.

During the failover process, they fail with the following error:
Errors: Missing gtid_executed system variable.

This is caused by an incorrect validation done by the
check_gtid_version method method of the Server class.

The method is wrongfully assuming that an empty result to the "SHOW
VARIABLES LIKE 'gtid_executed'" query is a signal that gtid_executed
variable does not exist in the system.

This patch removes that verification, since besides being wrong is
superfluous.  Ensuring that the server version is >= 5.6.9 and that
GTIDs are enabled are sufficient conditions to ensure that
gtid_executed variable will be on the system.  Furthermore it relaxes
the masks regarding server delay on the rpl_admin_gtid_loopback tests
to allow them to work on slower pcs.