BUG#22628187 : msyqldbcopy does not work with mysql 5.7
The mysqldbcopy utility is not working with MySQL 5.7. The utility
fails when trying to obtain foreign_key_checks system variable value
after locking the table it is trying to copy. This issue happens
because starting from MySQL 5.7.6 the the "session_variables" table
was moved from the INFORMATION_SCHEMA to the PERFORMANCE_SCHEMA, and
the tables in the INFORMATION_SCHEMA are the only exception of tables
than can be accessed without requiring to be specified by the LOCK
statement.

This patch addresses this issue by using SELECT @@GLOBAL.foreign_key_checks
instead, which avoid issues with table locking.