BUG#22781266 : EXCLUDE OPTION IN MYSQLDBEXPORT DOES NOT SEEM TO TAKE COMPLETE EFFECT
The REGEXP strings to identify the exclude patterns does not take into account strings
such as "db.tbl", "db.%", or even "%". This patch corrects the regex strings so that
these options are recognized and thus the user can now execute an export and exclude
tables that start with a set of characters such as "test%" and exclude all objects
whose name begins with "test".

But handling the special symbols requires a REGEXP directive \Z, which is only needed
for wildcard searches (where --exclude is used). A new parameter was added to append
the directive for searches where the % or * can appear at the end of the string.

The patch also corrects an omission where the --exclude option was not being sent to
the code in export. This is because the options dictionary was not being passed to the
class constructor.

v2