BUG#22934469: Allow MUT to run tests from a custom location
Problem
Specifying the --testdir option results in unexpected behavior when
verifying results because the path of the directory expected to contain
the test results is not modified accordingly. This option does not
provide support for execution of test suites and requires specifying
the t folder as part of the path. The working of --utildir option
needs to be verified.

Solution
RESULT_PATH and SUITE_PATH are now modified according to the value of
opt.testdir. The paths of each of the suites are appended to sys.path
to ensure execution of the test suites. The value of TEST_PATH is
also appended with 't' so that it is not needed to be specified.
The working of --utildir option has been verified, it is found to
work fine.