BUG#22076859 : AFTER PRESSING <DELETE> IN MYSQL UTILITY CONSOLE THE COMMAND BUFFER IS CORRUPT
This patch corrects a problem in mysqluc where the delete key press code
failed to position the cursor in the correct location. More specifically,
if the user deleted a character in the middle of the string, the code
reprints the remaining string but on moves the cursor back one position.

Thus, subsequent movement of the cursor with left, right or backspace or
delete causes the string to overlap and print ghost characters.

This patch adds code to "print" the correct number of backspaces to
reposition the cursor.

v2