BUG#21222808 : MYSQLDIFF: ENCODING COMMENTS ALWAYS MAKE DATABASES DEFINITIONS LOOK DIFFERENT
This patch removes the CREATE database from the diff output for mysqldiff and
mysqldbcompare for cases where the database name is the only difference. That
is, the difference output should not show a difference when comparing db1 to
db2 when the only difference is the name of the database.

However, the patch preserves change detection for differences in quote usage
and decorator changes (e.g. character set).