BUG#22878052 : ADD MASTER OUTAGE DELAY TO MYSQLFAILOVER
This patch adds a new option to mysqlfailover to allow the user to specify an
additional delay for detecting the master has failed. The new option allows
scenarios where the network may have issues or the server is busy or the server
is being rebooted, etc.

A new test was added to test the feature at the most basic level.

v2