BUG#22780854 : VALUEERROR WHILE USING MYSQLRPLSYNC DISCOVER-SLAVE OPTION WITH HOST:PORT
This patch fixes an issue with the parse_user_password method that incorrectly returns
a formatted error messages if there is a problem parsing <user,host>. The patch changes
it to raising an error instead. All uses of the method are now wrapped in a try block
with test cases added to show the errors are caught correcty. This affects --rpl-user
and --discover-slaves-login.

v2