BUG#17666275 : ADD AUTHENTICATION PLUGIN INFO TO MYSQLUSERCLONE
This patch corrects an issue when copying users that have been
created with the IDENTIFIED WITH authentication plugin option.
The utility would copy the user but not copy the authentication
plugin setting making it impossible to copy users with a plugin.

This patch corrects the problem by capturing the plugin name and
assigning it to the new user. The code will throw an error
if the plugin is not on the destination machine or is inactive.

A warning is issued if the user attempts to assign a password
to a new user when the source user has an authentication
plugin. In this case, the utility will use the password and not
the authentication plugin for the new user.

Tests added for functionality, warning, and error.

v3