BUG#22543517: MYSQLBINLOGPURGE DOES NOT WORK WITH OTHER THAN THE FIRST REPLICATION CHANNEL
This defect corrects a problem when running binlogpurge on a 5.7.6 or later replication
topology setup for multi-source replication. Specifically, when run against a master
on other than the first channel, binlogpurge will throw an error stating the slaves
cannot be discovered or that the slave is not connected to the master.

This patch corrects the problem by iterating through the slave status to check for
the requested master in the list. If not found, the original error(s) are generated
otherwise, the master is matched to one of the replication channels and progress
continues.

New tests added to show corrected behaviour.

CAB v2