BUG#23111808: MYSQLDBBCOPY COPY BLOB DATA WHEN SOME FIELD'S VALUE IS NULL
This patch addresses a problem when copying blob fields set to NOT NULL.
When blob data is copied, it is copied in two passes. The first pass
inserts all fields except BLOB fields. The second pass updates the rows
filling in the blob field data. This is the design of the copy and has
many advantages for bulk inserting.

However, when blob fields are marked NOT NULL, the copy process will fail
with an error about the field not being empty on the first pass. Since the
process does not tolerate blob fields set to NOT NULL, the code has been
changed to detect when this occurs and issue an error message with a work
around suggestion.

More specifically, this patch adds a check to the copy method to check all
tables in the list of databases for blob fields set to NOT NULL. If any
are found, the database and table are printed along with an error message
as described above.

CAB v2