BUG#22495193: MYSQLDBEXPORT DOES NOT DUMP BIT FIELDS CORRECTLY
This patch corrects an issue regarding bit fields and dbexport. The code was
not exporting data in the correct format for SQL output. Rather than displaying
the decimal value of the bit fields or b'???', the code was interpreting it
as a string, which produced incorrect results.

This patch corrects the problem by grouping the bit fields and using the
standard Python data type conversion for bit fields.

Test case added to verify correct behaviour.

CAB v2