BUG#22619159: SPECIAL CHAR IN PASSWORD CAUSES ERROR WITH MYSQLRPLSHOW
This patch corrects a problem using special characters in the --rpl-user option.
More specifically, if the user attempts to use a : in the password, the code fails
with an error stating there are too many values to unpack.

The problem is due to the use of split(':') in the code to read the user and
password from the option. Thus, the code was changed to split only on the first
: encountered (from the left). This allows the use of special characters in the
password (but not the user name).

This patch affects all utilities that use --rpl-user, not just the rplshow
utility. Comprehensive tests were run on the entire replication test suite as well
as a new test case added for passwords with special characters.

A note shall be placed in the documentation stating the limitation of using
special characters such as the colon in a user name for --rpl-user.