BUG#22932375: HA UTILTIES MUST SET CONNECTION_TIMEOUT IN C/PY CONNECTION TO SHORT VALUE
This patch adds the ability to specify a connection timeout value for C/Py for the
mysqlfailover utility. More specifically, it allows users to override the default
value of 10 for C/Py connection_timeout.

This patch only targets mysqlfailover since it is the only utility where this
feature would be beneficial and then only for certain use cases.

A test case was added to ensure there is no regression with existing features.