BUG#22905132 : HANDLING REGEXP FOR FEW OF THE UTILITIES
This patch adds checks for the --exclude option for the utilities
dbcompare, dbcopy, and dbexport. It also consolidates the --exclude
code to the options.py module.

The check will attempt to determine if the pattern supplied has any
special characters (such as an astericks), which may indicate that
the pattern could be a regex pattern. If there are special, non-SQL
LIKE pattern characters and the user has not specified the --regexp
option, a warning is presented to suggest the user check the pattern
for possible --regexp option.

Test cases were added to demonstrate and test the new behaviour.

Also, tests were updated with result file mismatches due to changes
in the server for InnoDB and triggers.