BUG#23300260 : MYSQL UTILITIES FAIL WITH "TOO MANY VALUES TO UNPACK" IN _UPDATE_ALIAS
This patch corrects an error that is possible for hostname aliases that have more
than (2) segments. For example, local.mydomain contains 2, localhost.mydomain.local
contains (3), which will result in a catastrophic error "too many values" causing
the utility to crash.

The error is corrected by limiting the results of the rsplit() call to at most (2)
elements.

A test case was added to is_alias.py that replicates the error.