BUG#24348569 : UTILITIES FAIL TO PROPERLY DETECT 5.7.13 SERVERS
This patch fixes an issue discovered for MySQL servers that have
a version string that includes more than 2 periods in the name.
For example, 5.7.13-0ubuntu0.16.04.2 for Linux on x86_64 ((Ubuntu)).
This string may result in a mismatch of the server version
causing serverclone and serverinfo to execute the wrong code
and ultimately produce warnings, errors, or fail.

The patch fixed the problem by detecting the number of periods
found and adjusting accordingly.

CAB v1