BUG#23335753 : mysqldbcopy issue with tinytext
When copying data that contains blob fields that are marked NOT NULL,
the utility will present an error and stop copying the data. The message
told the user to alter the columns to not include the nulls then reapply
the restriction after the copy.

This patch automates that procedure by capturing the column definition
before the copy and removing the NOT NULL restriction. After the copy,
the NOT NULL restriction is readded.

Thus, dbcopy can now copy tables that have blob fields marked as NOT NULL.

CAB v1