BUG#23598948 : tables with primary key of 0 with auto_increment fail to copy
When users attempt a copy of a table with a 0 value in an auto_increment column,
the utility will stop with an SQL error. This is normal behaviour for MySQL as
it does not allow 0 in auto_increment columns unless the SQL_MODE
NO_AUTO_VALUE_ON_ZERO is turned on (added to the sql_mode list).

This patch adds tools to check for auto_increment columns containing 0 as well
as a tool to turn a SQL_MODE on or off. These tools are used to change dbcopy
to allow copying of table data with auto_increment values of 0 by turning on
the SQL_MODE before the copy and turning it off after the copy.

For export, the user is presented with a warning and a sample SQL_MODE statement
is included in the output stream for the user to modify for the import utility.

Documentation shall be updated to explain the new policy for auto_increment
columns that contain 0.

CAB v2