BUG#22543116 : mysqlbinlogpurge requires the master port explicitly
This patch corrects the problem when socket files are used in
replication utilities. Previously, the port defaulted to 3306
instead of the actual port for the servers. This prevents the
use of some of the replication utilities with socket file
connections to the master and/or slaves specified on the command
line.

The code now retrieves the port from the connected servers so
that connections of slaves are directed to the correct port.

This is accomplished using a new method to retrieve the port
whenever a socket is used on Posix systems. This problem occurs
in several places in the source code. Thus, this patch also
consolidates and conforms the locations to work the same way.

CAB v2