BUG#22457829 : mysql serverclone doesn't detect the error properly and fails with ambiguous msg
Currently, the mysqlserverclone utility presents a cryptic and
difficult to understand error should the cloned server fail to
start.

This patch adds context to errors resulting from server startup
for mysqlserverclone. The user is directed to use the -vvv option
to find the actual error from the server startup sequence.

CAB v2