BUG#24657793 : MYSQLDBCOPY FAILS TO COPY VIEWS WHICH REFER OTHER VIEWS
Currently, the view dependency check fails to correctly identify all views
that depend on another view. The result is it is possible to construct a
set of views that can force mysqldbcopy to fail.

This patch improves the dependency check and captures any views that
fail to create due to dependency failures and reattempts to create them once
all of the views are created. While not 100% proof against all possible
dependencies, this solutions should catch the majority of common view
interdependencies.

One limitation is the solution does not take into account view
dependencies across databases or those not included in the database list
to be copied.

CAB v2