BUG#21201454 : mysqlfailover do not set read_only to slave during preparing candidate promotion
The code in topology.py does not correctly identify when the server
READ_ONLY variable is set. Thus, it is possible for READ_ONLY to not
be set during slave promotion for failover.

This patch corrects the logic allowing the candidate server to be
placed in READ_ONLY mode during the promotion. It also presents the
user with statements indicating the mode when run with verbosity
turned on.

CAB v2