BUG#20411237 : mysqlfrmreader must support new year data type
Currently, mysqlfrmreader in diagnostic mode does not correctly identify
the year datatype. In fact, it reads it as the older 2-byte type. Newer
servers use 4 bytes.

This patch improves the year datatype determination for newer servers
by correctly identifying the 4-byte version of the year datatype and
displays it as year(4).

CAB v2