BUG#18763591 : using grid format header difference should be ignored for mysqldbcompare
Currently, when using the GRID format for mysqldbcompare and there
are differences among the rows, the code duplicates the header when
the column sizes differ. This makes the output seem strange and
confusing and seems to identify the column names themselves as being
different.

This patch adds code to calculate the maximum column widths for
each table of the rows that differ using the maximum column values
in the formatted rows thereby removing the difference with the
column headers and extraneous print lines.

CAB v2