BUG#20857151 : MYSQLDIFF DROPS AND RECREATES PK AND INDEXES WHEN IT IS NOT NEEDED
This defect is subtle but nonetheless critical. The comparison of index metadata
includes a column for cardinality. This value is updated is sublte ways by the
server and can differ from one machine to another depending on how the data was
entered. The problem also creates a case where the compare_db and diff tests can fail
randomly producing (or not) differences with index drop and recreate.

The solution is to ignore cardinality when comparing index metadata. Once corrected,
the code no longer produces unnecessary index drop and recreate and makes the tests
stable. Cool.

CAB v1