BUG#21764003 : convenience notation not permitted
The mysqldiff utility permits the use of a single database when
comparing data on two servers, but the code does not work.

This patch corrects the behaviour by examining the database argument
looking for the colon (:) separator. If not found, it duplicates
the database name so that the code works properly.

CAB v3