BUG#24613794 : mysqlfrm unable to parse some .frm files
This patch corrects a problem found during diagnostic mode where .frm
files have indexes that exceed one page in the old page format. In this
case, the offset for the column headings must take into account the
size of the index keys in calculating the offset. This patch includes
the change along with additional test files and tests to verify the
changes.

CAB v4