BUG#24624716: ISSUE WITH 'MYSQLREPLICATE' 1.6.4 UTILITY
When connecting to a master and/or slave with the same user account as
the replication user and there exists another account with an anonymous
host, for example 'rpl'@'localhost' and 'rpl'@'%', where the anonymous
user account has full privileges, the mysqlreplicate utility can fail.

The failure is due to the way the code checks for privileges. Currently,
the code is designed to include the anonymous host user grants, but it
should not do so. The replication user account must use the more
restrictive user@host rather than an anonymous host.

Thus, this code has changed the way privileges are checked for the
replication user. It no longer uses the anonymous host and instead
checks only for the user@host as specified in the connection parameters.

This eliminates the reliance on anonymous host accounts as well as makes
the utility work correctly should a user desire to use the same account
for connecting to the master and the replication user.

CAB v2