BUG#25030151: mysqlfailover (v:1.6) was stopped silently when using a '--exec-before' script
External scripts executed with failover or switchover that fail are not
detected. In fact, the error is ignored. If the error is severe, the
utility can halt unexpectedly without error or messages of any kind.

This patch corrects the anomoly by capturing the error in an exception
and logging the error to the log. This keeps the original premise that
the external script should not halt the utility.

Also, this patch corrects a defect found in the failover test whereby
there was no coordination of the key phrases searched in the log for
detecting successful test cases. While unconfirmed, this may fix the
long-standing random failure seen on some platforms.

Finally, the patch introduces the ability to search for multiple key
phrases in the log for detecting successful test case execution for
failover tests.

CAB v2